[CS] Prepare for PHP-CS-Fixer 2.x

+++ lib/FieldValueMapper/PriceMapper.php
@@ -40,6 +40,6 @@ class PriceMapper extends FieldValueMapper
      */
     public function map(Field $field)
     {
-        return (double)$field->value;
+        return (float)$field->value;
     }
 }
